updating soft delete feature
