+++ README.md
@@ -294,6 +294,7 @@ These examples demonstrate common patterns and best practices for building conne
 - [three_operations](https://github.com/fivetran/fivetran_connector_sdk/tree/main/examples/common_patterns_for_connectors/three_operations) - This example shows how to use upsert, update and delete operations.
 - [tracking_tables](https://github.com/fivetran/fivetran_connector_sdk/tree/main/examples/common_patterns_for_connectors/tracking_tables) - This example shows how to track tables that have already been synced in order to be able add new tables and have them automatically sync back to an initial timestamp.
 - [update_configuration_during_sync](https://github.com/fivetran/fivetran_connector_sdk/tree/main/examples/common_patterns_for_connectors/update_configuration_during_sync) - This example shows how to update the configuration of the connector during a sync. It demonstrates how to modify the configuration values based on certain conditions using the Fivetran REST API.
+- [soft_delete](https://github.com/fivetran/fivetran_connector_sdk/tree/main/examples/common_patterns_for_connectors/soft_delete) - This example demonstrates Fivetran's soft delete feature using the truncate-and-reload pattern. On each sync, `op.truncate()` marks all existing warehouse rows as soft-deleted (`_fivetran_deleted = True`), then the full current snapshot is re-upserted. Rows absent from the new snapshot remain soft-deleted rather than physically removed.
 
 #### Workflows
